move generate proofs to helper
